New deploy

+++ bin/deploy.sh
@@ -7,27 +7,31 @@ export LOG_FILE="/tmp/bosh_deploy.log"
 
 set -e
 
-GEN_NEW_MANIFEST_FILE=0
-MANIFEST_FILE=${MANIFEST_FILE:-$WORKSPACE/bosh-solace-manifest.yml}
-INSTALL_BROKER=1
+export MANIFEST_FILE=${MANIFEST_FILE:-$WORKSPACE/bosh-solace-manifest.yml}
+GEN_NEW_MANIFEST_FILE=true
+INTERACTIVE=false
+INSTALL_BROKER=false
 
 CMD_NAME=`basename $0`
-BASIC_USAGE="usage: $CMD_NAME [-m MANIFEST_FILE][-c CI_CONFIG_FILE][-s][-h]"
+BASIC_USAGE="usage: $CMD_NAME [-m MANIFEST_FILE][-c CI_CONFIG_FILE][-i][-s][-h]"
 
 function showUsage() {
     read -r -d '\0' USAGE_DESCRIPTION << EOM
 $BASIC_USAGE
 
 Deploy BOSH VMRs.
-Omitting -m and -c will execute a basic bosh-manifest generator.
 
-Note: the -m and -c options cannot be used simultaneously.
+Default: A basic bosh-lite manifest will be generated and deployed with 1 instance of Shared-VMR
+
+Note 1: the -i option does nothing if -m or -c is given
+Note 2: the -m and -c options cannot be used simultaneously
 
 optional arguments:
   -m MANIFEST_FILE
         Manifest that will be deployed
   -c CI_CONFIG_FILE
         A Concourse property file from which a new bosh-manifest will be generated
+  -i    Will be prompted to interactively provide options to generate a bosh-lite manifest
   -s    Install the service broker after the deployment is finished
   -h    Show this help message and exit
 \0
@@ -35,22 +39,27 @@ EOM
     echo "$USAGE_DESCRIPTION"
 }
 
-while getopts :m:c:sh opt; do
+while getopts :m:c:ish opt; do
     case $opt in
         m)
             EXISTING_MANIFEST_FILE="$OPTARG"
             echo "Will use bosh-lite manifest file $EXISTING_MANIFEST_FILE"
-            cp $EXISTING_MANIFEST_FILE $MANIFEST_FILE
-            echo "Copied $EXISTING_MANIFEST_FILE to $MANIFEST_FILE"
+            if ! [ "$EXISTING_MANIFEST_FILE" -ef "$MANIFEST_FILE" ]; then
+                cp $EXISTING_MANIFEST_FILE $MANIFEST_FILE
+                echo "Copied $EXISTING_MANIFEST_FILE to $MANIFEST_FILE"
+            fi
             echo
-            GEN_NEW_MANIFEST_FILE=1;;
+            GEN_NEW_MANIFEST_FILE=false;;
         c)
             CI_CONFIG_FILE="$OPTARG"
-            echo "Will convert CI-config file, $OPTARG , to bosh-lite manifest file, $MANIFEST_FILE"
+            echo "Will convert CI-config file to bosh-lite manifest file:"
+            echo "    Input CI-Config:      $OPTARG"
+            echo "    Output Bosh Manifest: $MANIFEST_FILE"
             $SCRIPTPATH/parser/converter.py --in-file="$CI_CONFIG_FILE" --out-file="$MANIFEST_FILE"
             echo
-            GEN_NEW_MANIFEST_FILE=1;;
-        s) INSTALL_BROKER=0;;
+            GEN_NEW_MANIFEST_FILE=false;;
+        i)  INTERACTIVE=true;;
+        s)  INSTALL_BROKER=true;;
         h)
             showUsage
             exit 0;;
@@ -62,18 +71,24 @@ while getopts :m:c:sh opt; do
 done
 
 if [ -n "$EXISTING_MANIFEST_FILE" ] && [ -n "$CI_CONFIG_FILE" ]; then
+    showUsage
     >&2 echo "The -m and -c options cannot be used simultaneously"
     exit 1
 fi
 
-if [ "$GEN_NEW_MANIFEST_FILE" -eq "0" ]; then
+if $GEN_NEW_MANIFEST_FILE; then
     echo "A new manifest will be generated..."
     echo
-    $SCRIPTPATH/generateBoshManifest.py -h
-    echo
+    $INTERACTIVE && $SCRIPTPATH/generateBoshManifest.py -h && echo
     echo "MANIFEST_FILE set to $MANIFEST_FILE"
     echo
-    read -p "Please indicate the options that will be used to generate this manifest (Will proceed with the default settings if none provided): " MANIFEST_GEN_OPTS
+
+    if $INTERACTIVE; then
+        read -p "Please indicate the options that will be used to generate this manifest (Will proceed with the default settings if none provided): generateBoshManifest.py " MANIFEST_GEN_OPTS
+    else
+        echo "-i option was not given, manifest will be generated using default settings..."
+    fi
+
     echo
     $SCRIPTPATH/generateBoshManifest.py $MANIFEST_GEN_OPTS
     echo
@@ -83,10 +98,6 @@ $SCRIPTPATH/optimizeManifest.py $MANIFEST_FILE
 echo
 $SCRIPTPATH/deployBoshManifest.sh $MANIFEST_FILE
 echo
-
-if [ "$INSTALL_BROKER" -eq "0" ]; then
-    $SCRIPTPATH/installServiceBroker.sh
-fi
-
+$INSTALL_BROKER && $SCRIPTPATH/installServiceBroker.sh
 echo
 $SCRIPTPATH/updateServiceBrokerAppEnvironment.sh
